# docs/swagger.yaml
definitions:
...
  types.Event:
    properties:
      area:
        type: string
      boot_time:
        type: string
      client:
        type: string
      content:
        type: string
      event_id:
        type: string
      facility:
        type: integer
      host_address:
        type: string
      host_label:
        items:
          type: string
        type: array
      host_uuid:
        type: string
      hostname:
        type: string
      location:
        type: string
      mac_address:
        items:
          type: string
        type: array
      machine_id:
        type: string
      org:
        type: string
      org_group:
        type: string
      priority:
        type: integer
      severity:
        type: integer
      tag:
        type: string
      time:
        type: string
      tls_peer:
        type: string
    type: object
  types.Events:
    properties:
      events:
        items:
          $ref: '#/definitions/types.Event'
        type: array
    type: object
...
info:
  contact:
    email: skipper@southwinds.io
    url: http://www.southwinds.io/
  description: Control Service for Pilot agents
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Pilot Control Service
  version: 1.0.0
paths:
...
  /events:
    get:
      description: Returns a list of syslog entries following the specified filter
      parameters:
      - description: the host UUID of the entries to retrieve
        in: query
        name: uuid
        type: string
      - description: the organisation of the device where the syslog entry was created
        in: query
        name: og
        type: string
      - description: the organisation of the device where the syslog entry was created
        in: query
        name: or
        type: string
      - description: the area of the device where the syslog entry was created
        in: query
        name: ar
        type: string
      - description: the location of the device where the syslog entry was created
        in: query
        name: lo
        type: string
      - description: syslog entry tag
        in: query
        name: tag
        type: string
      - description: the syslog entry priority
        in: query
        name: pri
        type: string
      - description: the syslog entry severity
        in: query
        name: sev
        type: string
      - description: the time FROM which syslog entries are shown (time format must
          be ddMMyyyyHHmmSS)
        in: query
        name: from
        type: string
      - description: the time TO which syslog entries are shown (time format must
          be ddMMyyyyHHmmSS)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get filtered events
      tags:
      - Query
    post:
      description: submits syslog events to be persisted for further use
      parameters:
      - description: the events to submit
        in: body
        name: command
        required: true
        schema:
          $ref: '#/definitions/types.Events'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Submit Syslog Events
      tags:
      - Receiver
...
swagger: "2.0"